requirements.txt
flask
flask-cors
requests
sentence-transformers
faiss-cpu